Launching History activity Moved HistoryList fragment
