+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityLauncher.java
@@ -33,6 +33,7 @@ import org.wordpress.android.ui.accounts.SiteCreationActivity;
 import org.wordpress.android.ui.activitylog.detail.ActivityLogDetailActivity;
 import org.wordpress.android.ui.activitylog.list.ActivityLogListActivity;
 import org.wordpress.android.ui.comments.CommentsActivity;
+import org.wordpress.android.ui.history.HistoryActivity;
 import org.wordpress.android.ui.quickstart.QuickStartActivity;
 import org.wordpress.android.ui.main.SitePickerActivity;
 import org.wordpress.android.ui.main.WPMainActivity;
@@ -412,6 +413,13 @@ public class ActivityLauncher {
         intent.putExtra(EditPostActivity.EXTRA_IS_PAGE, true);
         intent.putExtra(EditPostActivity.EXTRA_IS_PROMO, false);
         fragment.startActivityForResult(intent, RequestCodes.EDIT_POST);
+    }
+
+    public static void viewPostOrPageHistoryForResult(Activity activity, SiteModel site, PostModel post) {
+        Intent intent = new Intent(activity, HistoryActivity.class);
+        intent.putExtra(WordPress.SITE, site);
+        intent.putExtra(EditPostActivity.EXTRA_POST_LOCAL_ID, post.getId());
+        activity.startActivityForResult(intent, RequestCodes.VIEW_HISTORY);
     }
 
     /*
