+++ WordPress/src/main/java/org/wordpress/android/ui/RequestCodes.java
@@ -25,6 +25,7 @@ public class RequestCodes {
     public static final int ACTIVITY_LOG_DETAIL = 1700;
     public static final int PAGE_PARENT = 1800;
     public static final int QUICK_START = 1800;
+    public static final int VIEW_HISTORY = 1900;
 
     // Media
     public static final int PICTURE_LIBRARY = 2000;
